feat: pass html-webpack-plugin options to config builder
Allows users to pass an additional, optional, argument to `html()` which
contains plugin options for `html-webpack-plugin`, enabling deeper
customisation.

Also fixes the `full-site` example test as newer versions of jest don't
include the `fail()` function, and we no longer need to test for
`-webkit-` vendor prefixes in the CSS output.